Separate guide and workshop material
Content is following a standard arrangement first implemented in the
dcoders/intro-web repository.

Guide material will be in the reference directory and workshop material
will be in the workshop directory.

In addition, the README will contain a brief summary/landing page for
members to read and understand the purpose for each section
(guide/workshop).